pyproject.toml
[project]
name = "Melchior"
version = "0.0.1"
description = ""
authors = [
    { name = "nullqwertyuiop", email = "null@member.fsf.org" },
]
dependencies = [
    "ichika>=0.0.6",
    "graiax-shortcut>=0.2.1",
    "graia-broadcast>=0.19.2",
    "creart-graia>=0.1.5",
    "kayaku>=0.5.2",
    "pydantic>=1.10.7",
    "graiax-fastapi>=0.2.1",
]
requires-python = ">=3.10,<4.0"
license = { text = "AGPL-3.0 license" }

[tool.pdm]

[tool.pdm.dev-dependencies]
dev = [
    "pre-commit>=3.3.1",
    "black>=23.3.0",
]

[tool.pdm.build]
includes = []

[tool.isort]
profile = "black"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
